src/api/requirements.txt
Django==4.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
django-cors-headers==3.13.0
Pillow==9.2.0